project restructuring improvements
