--- apps/game-engine/src/events/index.ts
@@ -1,2 +1,0 @@
-
-export * from './handlers';
