+++ apps/game-engine/src/triggers/activity-creation.trigger.ts
@@ -1,10 +1,10 @@
-import { CreateActivityInput, UserActivity, WebhookEvent, logger } from '@codeheroes/common';
+import { CreateActivityInput, WebhookEvent, logger } from '@codeheroes/common';
 import { onDocumentCreated } from 'firebase-functions/v2/firestore';
 import { DatabaseService } from '../core/services/database.service';
-import { XpCalculatorService } from '../core/services/xp-calculator.service';
-import { EventUtils } from './event.utils';
+import { EventUtils } from '../core/utils/event.utils';
 
-export const createUserActivity = onDocumentCreated('events/{eventId}', async (event) => {
+
+export const activityCreationTrigger = onDocumentCreated('events/{eventId}', async (event) => {
   logger.info('New event document created', {
     eventId: event.params.eventId,
   });
@@ -41,32 +41,3 @@ export const createUserActivity = onDocumentCreated('events/{eventId}', async (e
 
   await dbService.createUserActivity(activityInput);
 });
-
-export const calculateUserXp = onDocumentCreated(
-  'users/{userId}/activities/{activityId}',
-  async (event) => {
-    const activity = event.data?.data() as UserActivity;
-    
-    if (!activity) {
-      logger.error('No activity data found');
-      return;
-    }
-
-    const xpCalculator = new XpCalculatorService();
-    const xpResult = xpCalculator.calculateXp(activity);
-
-    logger.info('Calculated XP for activity', {
-      userId: event.params.userId,
-      activityId: event.params.activityId,
-      ...xpResult,
-    });
-
-    const dbService = new DatabaseService();
-    await dbService.updateUserXp(
-      event.params.userId,
-      event.params.activityId,
-      xpResult,
-      activity
-    );
-  }
-);
