+++ apps/game-engine/src/triggers/index.ts
@@ -1,0 +1,3 @@
+export * from './activity-creation.trigger';
+export * from './xp-calculation.trigger';
+
